+++ changelog.md
@@ -7,7 +7,23 @@
     * Added `minmax`
     * Added `unitlength`
     * Scalers can be fitted and mapped through a helper type - this is for fitting to a training set then applying to a test set
-- utils/eval.jl
-    * Added ROCStats and functions
+- utils/classification_eval.jl
+    * New type: ClassificationStatistics
+    * New type: ConfusionMatrix
     * Added many more classification stats
-    * Confusion matrix
+        * tp, fp, tn, fn calculators
+        * positive_predictive_value (precision)
+        * negative_predictive_value
+        * sensitivity (true positive rate)
+        * specificity (true negative rate)
+        * false_positive_rate
+        * false_negative_rate
+        * false_discovery_rate
+        * matthews_corrcoef
+        * informedness
+        * markedness
+        * f1score
+        * accuracy (existing, listing it here for a complete list)
+    * Receiver Operating Characteristics
+- utils/utils.jl
+    * `classify`